+++ frontend/src/lib/components/neurons/NeuronsTitle.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-  import Summary from "$lib/components/nav/Summary.svelte";
-  import PrincipalText from "$lib/components/nav/PrincipalText.svelte";
+  import Summary from "$lib/components/summary/Summary.svelte";
+  import PrincipalText from "$lib/components/summary/PrincipalText.svelte";
 </script>
 
 <Summary>
