+++ frontend/src/lib/components/summary/PrincipalText.svelte
