+++ frontend/src/lib/components/summary/ProjectLogo.svelte
